finally making headway.
I have all of the gets working for fan speed, core voltage, core clock and memory clock

I also sorted out the issue with it spitting out a huge number of displays. I still have to code in proper handling of the displays but I do have most of the ports sorted out correctly. I will not be able to add in the multiport active dongles until I have someone use the library that has one so I can see how the data is set up.

It is shaping up pretty good so far. still have a bunch more to add in like display adjustments and the eyefinity crap. I have the code added for crossfire. I am not able to test this either because I do not have 2 cards that I can crossfire together.

I still have to add in overdrive 6, 7, and 8. I am trying to make a common interface for all of them. That is going to take some work to accomplish.

I added an example file which shows how to access some of the various bits and pieces that are currently working. at the present moment the code is good for displaying and not setting anything as of yet. I think I have gotten the hang of the poor documentation for the ADL and the odd way in which it works.